refactor: GameLogic to BotService, DisplayService and PlayerService.
