+++ model/src/main/java/entity/Tile.java
@@ -3,22 +3,12 @@ package entity;
 import enums.Color;
 
 public class Tile {
-    private Color color;
     private int x;
     private int y;
 
-    public Tile(Color color, int x, int y) {
-        this.color = color;
+    public Tile(int x, int y) {
         this.x = x;
         this.y = y;
-    }
-
-    public enums.Color getColor() {
-        return color;
-    }
-
-    public void setColor(Color color) {
-        this.color = color;
     }
 
     public int getX() {
