+++ services/src/main/java/repository/ReversiListener.java
@@ -1,9 +1,14 @@
 package repository;
 
+import entity.Board;
+import entity.Player;
+import enums.Color;
 import enums.GameStatus;
 
+import java.util.Map;
+
 public interface ReversiListener {
-    boolean moveMade(long playerId, int x, int y);
+    boolean moveMade(Board board, Map<String, Player> players, String currentPlayerId);
     boolean moveSkipped(long playerId);
 
     GameStatus gameFinished();
@@ -13,9 +18,10 @@ public interface ReversiListener {
 
     boolean scoreUpdated();
 
-    byte playerTurn();
+    String playerTurn(String currentPlayerId, Map<String, Player> players);
 
-    boolean playerJoin(long playerId);
+    boolean playerJoin(Map<String, Player> players, String id, Color color);
+    boolean botJoin(Map<String, Player> players, String id, Color color);
     boolean playerLeave(long playerId);
     boolean playerDisconnect(long playerId);
